separate ContributorUtils to RectorGenerator and DocumentationGenerator
