+++ src/Console/Application.php
@@ -6,9 +6,9 @@ use Jean85\PrettyVersions;
 use Rector\Configuration\Configuration;
 use Rector\Console\Output\JsonOutputFormatter;
 use Rector\Exception\Configuration\InvalidConfigurationException;
-use Rector\Utils\ContributorTools\Command\DumpNodesCommand;
-use Rector\Utils\ContributorTools\Command\DumpRectorsCommand;
-use Rector\Utils\ContributorTools\Exception\Command\ContributorCommandInterface;
+use Rector\Utils\DocumentationGenerator\Command\DumpNodesCommand;
+use Rector\Utils\DocumentationGenerator\Command\DumpRectorsCommand;
+use Rector\Utils\RectorGenerator\Contract\ContributorCommandInterface;
 use RuntimeException;
 use Symfony\Component\Console\Application as SymfonyApplication;
 use Symfony\Component\Console\Command\Command;
