--- utils/ContributorTools/src/Exception/Command/ContributorCommandInterface.php
@@ -1,7 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Rector\Utils\ContributorTools\Exception\Command;
-
-interface ContributorCommandInterface
-{
-}
