--- utils/ContributorTools/src/Exception/ConfigurationException.php
@@ -1,9 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Rector\Utils\ContributorTools\Exception;
-
-use Exception;
-
-final class ConfigurationException extends Exception
-{
-}
