+++ utils/DocumentationGenerator/config/config.yaml
@@ -1,0 +1,9 @@
+services:
+    _defaults:
+        public: true
+        autowire: true
+        autoconfigure: true
+
+    Rector\Utils\DocumentationGenerator\:
+        resource: '../src'
+        exclude: '../src/{Exception,Configuration/Configuration.php,Node/*Info.php}'
