+++ utils/DocumentationGenerator/src/Contract/OutputFormatter/DumpRectorsOutputFormatterInterface.php
@@ -1,0 +1,16 @@
+<?php declare(strict_types=1);
+
+namespace Rector\Utils\DocumentationGenerator\Contract\OutputFormatter;
+
+use Rector\Contract\Rector\RectorInterface;
+
+interface DumpRectorsOutputFormatterInterface
+{
+    public function getName(): string;
+
+    /**
+     * @param RectorInterface[] $genericRectors
+     * @param RectorInterface[] $packageRectors
+     */
+    public function format(array $genericRectors, array $packageRectors): void;
+}
