+++ utils/RectorGenerator/config/config.yaml
@@ -1,0 +1,9 @@
+services:
+    _defaults:
+        public: true
+        autowire: true
+        autoconfigure: true
+
+    Rector\Utils\RectorGenerator\:
+        resource: '../src'
+        exclude: '../src/{Exception,Configuration/Configuration.php}'
