+++ utils/RectorGenerator/src/Contract/ContributorCommandInterface.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Rector\Utils\RectorGenerator\Contract;
+
+interface ContributorCommandInterface
+{
+}
