+++ utils/RectorGenerator/src/Exception/ConfigurationException.php
@@ -1,0 +1,9 @@
+<?php declare(strict_types=1);
+
+namespace Rector\Utils\RectorGenerator\Exception;
+
+use Exception;
+
+final class ConfigurationException extends Exception
+{
+}
