+++ utils/RectorGenerator/templates/packages/_Package_/src/Rector/_Category_/_Name_.php.inc
