+++ utils/RectorGenerator/templates/packages/_Package_/tests/Rector/_Category_/_Name_/Fixture/fixture.php.inc
